backend/requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
django-rest-auth==0.9.5
django-cors-headers==3.5.0
sqlparse==0.4.1
pytz==2020.1
psycopg2-binary==2.8.6
joblib==0.16.0
gunicorn==20.0.4